Fix running issues for Linux
It is not a problem to run scripts without executable permission via Git
Bash on Windows but it is impossible to do the same on Linux shell.
Additionally, when file or command is not in */bin/ directory, you cannot
run it in current directory providing only its name. This patch fixes
these issues making scripts executable and adding './' before test name
via run-tests script.